+++ src/eppasm.cpp
@@ -159,6 +159,10 @@ extern "C" {
     int *med_cd4init_cat = INTEGER(getListElement(s_fp, "med_cd4init_cat"));
     int *med_cd4init_input = INTEGER(getListElement(s_fp, "med_cd4init_input"));
 
+    int art_alloc_method = *INTEGER(getListElement(s_fp, "art_alloc_method"));
+    double art_alloc_mxweight = *REAL(getListElement(s_fp, "art_alloc_mxweight"));
+    int scale_cd4_mort = *INTEGER(getListElement(s_fp, "scale_cd4_mort"));
+
 
     // diagnosis model
     int t_hts_start = *INTEGER(getListElement(s_fp, "t_hts_start")) - 1; // -1 for 0-based indexing in C vs. 1-based in R
@@ -661,6 +665,9 @@ extern "C" {
       ////  HIV model simulation  ////
       ////////////////////////////////
 
+      int cd4elig_idx = artcd4elig_idx[t] - 1; // -1 for 0-based indexing vs. 1-based in R
+      int anyelig_idx = (specpop_percelig[t] > 0 | pw_artelig[t] > 0) ? 0 : (who34percelig > 0) ? hIDX_CD4_350 : cd4elig_idx;
+
       for(int hts = 0; hts < HIVSTEPS_PER_YEAR; hts++){
 
         int ts = (t-1)*HIVSTEPS_PER_YEAR + hts;
@@ -675,7 +682,16 @@ extern "C" {
         for(int g = 0; g < NG; g++)
           for(int ha = 0; ha < hAG; ha++){
             for(int hm = 0; hm < hDS; hm++){
-              double deaths = cd4_mort[g][ha][hm] * hivpop[t][g][ha][hm];
+
+	      double cd4mx_scale = 1.0;
+	      if(scale_cd4_mort & t >= t_ART_start & hm >= anyelig_idx){
+		double artpop_hahm = 0.0;
+		for(int hu = 0; hu < hTS; hu++)
+		  artpop_hahm += artpop[t][g][ha][hm][hu];
+		cd4mx_scale = hivpop[t][g][ha][hm] / (hivpop[t][g][ha][hm] + artpop_hahm);
+	      }
+	      
+              double deaths = cd4mx_scale * cd4_mort[g][ha][hm] * hivpop[t][g][ha][hm];
               hivdeaths_ha[g][ha] += DT*deaths;
 	      hivpopdeaths[t][g][ha][hm] += DT*deaths;
               grad[g][ha][hm] = -deaths;
@@ -809,8 +825,6 @@ extern "C" {
 	
         // ART progression, mortality, and initiation
         if(t >= t_ART_start){
-          int cd4elig_idx = artcd4elig_idx[t] - 1; // -1 for 0-based indexing vs. 1-based in R
-          int anyelig_idx = (specpop_percelig[t] > 0 | pw_artelig[t] > 0) ? 0 : (who34percelig > 0) ? hIDX_CD4_350 : cd4elig_idx;
 
           // progression and mortality
           for(int g = 0; g < NG; g++)
@@ -871,7 +885,7 @@ extern "C" {
                     frr_pop_ha += frr_art[t][ha-hIDX_FERT][hm][hu] * artpop[t][g][ha][hm][hu];
                 }
                 for(int hm = anyelig_idx; hm < cd4elig_idx; hm++){
-                  double pw_elig_hahm = DT * births_by_ha[ha-hIDX_FERT] * frr_cd4[t][ha-hIDX_FERT][hm] * hivpop[t][g][ha][hm] / frr_pop_ha;
+                  double pw_elig_hahm = births_by_ha[ha-hIDX_FERT] * frr_cd4[t][ha-hIDX_FERT][hm] * hivpop[t][g][ha][hm] / frr_pop_ha;
                   artelig_hahm[ha-hIDX_15PLUS][hm] += pw_elig_hahm;
                   Xartelig_15plus += pw_elig_hahm;
                   expect_mort_artelig15plus += cd4_mort[g][ha][hm] * pw_elig_hahm;
@@ -991,7 +1005,32 @@ extern "C" {
 
 	      } // end loop over ha
 
-            } else { // Use mixture of eligibility and expected mortality for initiation distribution
+            } else if(art_alloc_method == 4) {  // lowest CD4 first
+
+	      for(int hm = hDS-1; hm >= anyelig_idx; hm--){
+		double artelig_hm = 0;
+		for(int ha = hIDX_15PLUS; ha < hAG; ha++)
+		  artelig_hm += artelig_hahm[ha-hIDX_15PLUS][hm];
+		double init_prop = (artelig_hm == 0 | artinit_hts > artelig_hm) ? 1.0 : artinit_hts / artelig_hm;
+
+		for(int ha = hIDX_15PLUS; ha < hAG; ha++){
+		  double artinit_hahm = init_prop * artelig_hahm[ha-hIDX_15PLUS][hm];
+
+		  if(t >= t_hts_start){
+		    Rprintf("Lowest CD4 first ART allocation option not yet implemented for HIV testing model\n. STOP WHAT YOU ARE DOING NOW");
+		    break;
+		  }
+		  
+		  hivpop[t][g][ha][hm] -= artinit_hahm;
+                  artpop[t][g][ha][hm][ART0MOS] += artinit_hahm;
+		  artinits[t][g][ha][hm] += artinit_hahm;
+		}
+		if(init_prop < 1.0)
+		  break;
+		artinit_hts -= init_prop * artelig_hm;
+	      }
+	      
+	    } else { // Use mixture of eligibility and expected mortality for initiation distribution
 
               for(int ha = hIDX_15PLUS; ha < hAG; ha++) {
 
@@ -1002,7 +1041,8 @@ extern "C" {
 		    undiagnosed_ha += hivpop[t][g][ha][hm] - diagnpop[t][g][ha][hm];
 		
                 for(int hm = anyelig_idx; hm < hDS; hm++){
-                  double artinit_hahm = artinit_hts * artelig_hahm[ha-hIDX_15PLUS][hm] * 0.5 * (1.0/Xartelig_15plus + cd4_mort[g][ha][hm] / expect_mort_artelig15plus);
+
+                  double artinit_hahm = artinit_hts * artelig_hahm[ha-hIDX_15PLUS][hm] * ((1.0 - art_alloc_mxweight)/Xartelig_15plus + art_alloc_mxweight * cd4_mort[g][ha][hm] / expect_mort_artelig15plus);
                   if(artinit_hahm > artelig_hahm[ha-hIDX_15PLUS][hm])
 		    artinit_hahm = artelig_hahm[ha-hIDX_15PLUS][hm];